+++ pytorch_lightning/core/lightning.py
@@ -1,7 +1,6 @@
 import collections
 import inspect
 import os
-import warnings
 from abc import ABC, abstractmethod
 from argparse import Namespace
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union, Sequence
@@ -20,7 +19,7 @@ from pytorch_lightning.core.memory import ModelSummary
 from pytorch_lightning.core.saving import ModelIO, load_hparams_from_tags_csv
 from pytorch_lightning.overrides.data_parallel import LightningDistributedDataParallel
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities.warnings import RankWarning
+from pytorch_lightning.utilities import rank_zero_warn
 
 try:
     import torch_xla.core.xla_model as xm
@@ -34,10 +33,6 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-
-        self._rank = 0
-        #: private warning stream
-        self._warning = RankWarning()
 
         #: Current dtype
         self.dtype = torch.FloatTensor
@@ -76,17 +71,6 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
         self.use_amp = False
 
         self.hparams = None
-
-    @property
-    def rank(self) -> int:
-        """Process rank. In general, metrics should only be logged by the process with rank 0."""
-        return self._rank
-
-    @rank.setter
-    def rank(self, value: int) -> None:
-        """Set the process rank."""
-        self._rank = value
-        self._warning.rank = value
 
     def print(self, *args, **kwargs) -> None:
         r"""
@@ -241,7 +225,8 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
             The loss value shown in the progress bar is smoothed (averaged) over the last values,
             so it differs from the actual loss returned in train/validation step.
         """
-        self._warning('`training_step` must be implemented to be used with the Lightning Trainer')
+        rank_zero_warn('`training_step` must be implemented to be used with the Lightning Trainer',
+                       trainer=self.trainer)
 
     def training_end(self, *args, **kwargs):
         """
@@ -1104,7 +1089,8 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
                   }
 
         """
-        self._warning('`configure_optimizers` must be implemented to be used with the Lightning Trainer')
+        rank_zero_warn('`configure_optimizers` must be implemented to be used with the Lightning Trainer',
+                       trainer=self.trainer)
 
     def optimizer_step(
             self,
@@ -1307,7 +1293,8 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
                     return loader
 
         """
-        self._warning('`train_dataloader` must be implemented to be used with the Lightning Trainer')
+        rank_zero_warn('`train_dataloader` must be implemented to be used with the Lightning Trainer',
+                       trainer=self.trainer)
 
     def tng_dataloader(self):  # todo: remove in v1.0.0
         """
@@ -1315,8 +1302,9 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
             Deprecated in v0.5.0. Use :meth:`train_dataloader` instead. Will be removed in 1.0.0.
         """
         output = self.train_dataloader()
-        self._warning("`tng_dataloader` has been renamed to `train_dataloader` since v0.5.0."
-                      " and this method will be removed in v1.0.0", DeprecationWarning)
+        rank_zero_warn("`tng_dataloader` has been renamed to `train_dataloader` since v0.5.0."
+                       " and this method will be removed in v1.0.0", DeprecationWarning,
+                       trainer=self.trainer)
         return output
 
     def test_dataloader(self) -> Union[DataLoader, List[DataLoader]]:
@@ -1423,7 +1411,7 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
             Deprecated in version 0.7.0. You should use :meth:`load_from_checkpoint` instead.
             Will be removed in v0.9.0.
         """
-        warnings.warn(
+        rank_zero_warn(
             "`load_from_metrics` method has been unified with `load_from_checkpoint` in v0.7.0."
             " The deprecated method will be removed in v0.9.0.", DeprecationWarning
         )
@@ -1535,7 +1523,7 @@ class LightningModule(ABC, GradInformation, ModelIO, ModelHooks):
                 is_namespace = checkpoint.get('hparams_type', 'namespace') == 'namespace'
                 hparams = Namespace(**ckpt_hparams) if is_namespace else ckpt_hparams
             else:
-                warnings.warn(
+                rank_zero_warn(
                     f"Checkpoint does not contain hyperparameters but {cls.__name__}'s __init__ "
                     f"contains argument 'hparams'. Will pass in an empty Namespace instead."
                     " Did you forget to store your model hyperparameters in self.hparams?"