+++ pytorch_lightning/logging/mlflow_logger.py
@@ -3,9 +3,9 @@
  The deprecated module name will be removed in v0.8.0.
 """
 
-import warnings
+from pytorch_lightning.utilities import rank_zero_warn
 
-warnings.warn("`mlflow_logger` module has been renamed to `mlflow` since v0.6.0."
-              " The deprecated module name will be removed in v0.8.0.", DeprecationWarning)
+rank_zero_warn("`mlflow_logger` module has been renamed to `mlflow` since v0.6.0."
+               " The deprecated module name will be removed in v0.8.0.", DeprecationWarning)
 
 from pytorch_lightning.loggers.mlflow import MLFlowLogger  # noqa: E402