+++ pytorch_lightning/utilities/__init__.py
@@ -1,0 +1,3 @@
+"""General utilities"""
+
+from pytorch_lightning.utilities.warnings import rank_zero_warn